+++ server/internal/storage/user.go
@@ -6,7 +6,7 @@ import (
 )
 
 type UserStorage interface {
-	Create(ctx context.Context, user *model.User) (int, error)
-	GetById(ctx context.Context, id int) (*model.User, error)
+	Create(ctx context.Context, user *model.User) (int64, error)
+	GetById(ctx context.Context, id int64) (*model.User, error)
 	GetByEmail(ctx context.Context, email string) (*model.User, error)
 }
